Pass mapping sizes in uint32_t instead of size_t to avoid warnings. (PGMR3MapPT, PGMR3MappingsSize and PGMR3MappingsFix.)
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@6915 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/VMM/PGM.cpp
@@ -3884,7 +3884,7 @@ PGMR3DECL(int) PGMR3ChangeShwPDMappings(PVM pVM, bool fEnable)
 {
     pVM->pgm.s.fDisableMappings = !fEnable;
 
-    size_t cb;
+    uint32_t cb;
     int rc = PGMR3MappingsSize(pVM, &cb);
     AssertRCReturn(rc, rc);
